pawswiipe_web/requirements.txt
Django==4.2.5
mysqlclient==2.1.0
channels==4.0.0
channels-redis==4.2.0
django-cors-headers==4.3.1
django-model-utils==4.4.0
django-private-chat2==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
redis==5.0.1
requests>=2.25.1
Pillow>=8.2.0

